Moved log out button from edit profile to nav bar
